requirements.txt
numpy
pandas
scipy
scikit-learn
xgboost
rdkit-pypi
shap
matplotlib
seaborn
mordred
